+++ src/register/fpscr.rs
@@ -1,5 +1,7 @@
 //! Floating-point Status Control Register
 
+use core::arch::asm;
+
 /// Floating-point Status Control Register
 #[derive(Clone, Copy, Debug)]
 pub struct Fpscr {
@@ -293,7 +295,8 @@ impl RMode {
 /// Read the FPSCR register
 #[inline]
 pub fn read() -> Fpscr {
-    let r: u32 = call_asm!(__fpscr_r() -> u32);
+    let r;
+    unsafe { asm!("vmrs {}, fpscr", out(reg) r, options(nomem, nostack, preserves_flags)) };
     Fpscr::from_bits(r)
 }
 
@@ -301,5 +304,5 @@ pub fn read() -> Fpscr {
 #[inline]
 pub unsafe fn write(fpscr: Fpscr) {
     let fpscr = fpscr.bits();
-    call_asm!(__fpscr_w(fpscr: u32));
+    asm!("vmsr fpscr, {}", in(reg) fpscr, options(nomem, nostack));
 }