+++ src/register/lr.rs
@@ -1,17 +1,20 @@
 //! Link register
 
+#[cfg(cortex_m)]
+use core::arch::asm;
+
 /// Reads the CPU register
-///
-/// **NOTE** This function is available if `cortex-m` is built with the `"inline-asm"` feature.
+#[cfg(cortex_m)]
 #[inline]
 pub fn read() -> u32 {
-    call_asm!(__lr_r() -> u32)
+    let r;
+    unsafe { asm!("mov {}, lr", out(reg) r, options(nomem, nostack, preserves_flags)) };
+    r
 }
 
 /// Writes `bits` to the CPU register
-///
-/// **NOTE** This function is available if `cortex-m` is built with the `"inline-asm"` feature.
+#[cfg(cortex_m)]
 #[inline]
 pub unsafe fn write(bits: u32) {
-    call_asm!(__lr_w(bits: u32));
+    asm!("mov lr, {}", in(reg) bits, options(nomem, nostack, preserves_flags));
 }