+++ src/register/msplim.rs
@@ -1,13 +1,17 @@
 //! Main Stack Pointer Limit Register
 
+use core::arch::asm;
+
 /// Reads the CPU register
 #[inline]
 pub fn read() -> u32 {
-    call_asm!(__msplim_r() -> u32)
+    let r;
+    unsafe { asm!("mrs {}, MSPLIM", out(reg) r, options(nomem, nostack, preserves_flags)) };
+    r
 }
 
 /// Writes `bits` to the CPU register
 #[inline]
 pub unsafe fn write(bits: u32) {
-    call_asm!(__msplim_w(bits: u32))
+    asm!("msr MSPLIM, {}", in(reg) bits, options(nomem, nostack, preserves_flags));
 }